analytics/requirements.txt
tweepy
python-dotenv
openai
flask